game-scrapper/python3/requirements.txt
requests==2.26.0
psycopg2-binary==2.9.1
beautifulsoup4==4.10.0
